+++ new_curse.h
@@ -209,7 +209,6 @@ extern void wclrtoeol P_((WINDOW *window));
 extern void wrefresh P_((WINDOW *window));
 extern void touchwin P_((WINDOW *window));
 extern void wnoutrefresh P_((WINDOW *window));
-extern void flushinp P_((void));
 extern void ungetch P_((int c));
 extern int wgetch P_((WINDOW *window));
 extern void Clear P_((int));
@@ -228,9 +227,7 @@ extern void raw P_((void));
 extern void noraw P_((void));
 extern void nl P_((void));
 extern void nonl P_((void));
-extern void saveterm P_((void));
 extern void fixterm P_((void));
-extern void resetterm P_((void));
 extern void nodelay P_((WINDOW *window, int flag));
 extern void idlok P_((WINDOW *window, int flag));
 extern void keypad P_((WINDOW *window, int flag));
